+++ main.py
@@ -1,16 +1,17 @@
 import os
 from argparse import ArgumentParser
+import logging as log
 from xml.dom.minidom import parse
 from mido import Message, MetaMessage, MidiFile, MidiTrack
 from predictGeneralMidi import getGeneralMidiNumber
 
-# Get Ardour file path
+# Process command line arguments
 parser = ArgumentParser()
 parser.add_argument("-f", "--file", dest="filename",
                     help="Optionally include the Ardour source file from the command line. Use Linux-style path conventions (like  ~/Ardour Folder/My Session.ardour) or Windows (like C:/Users/username/Ardour Folder/My Session.ardour).", metavar="FILE")
-# parser.add_argument("-q", "--quiet",
-#                     action="store_false", dest="verbose", default=True,
-#                     help="don't print status messages to stdout")
+parser.add_argument("-v", "--verbose",
+                    action="store_true", dest="verbose", default=False,
+                    help="Display MIDI messages and other debugging information.")
 
 args = parser.parse_args()
 
@@ -19,6 +20,12 @@ if not ardourFile:
     print("Please enter the path to your Ardour file. (E.g. ~/Ardour Folder/My Session.ardour ")
     ardourFile = os.path.abspath(os.path.expanduser(input(": ")))
 
+if args.verbose:
+    def vprint(*args, **kwargs):
+        print(*args, **kwargs)
+else:
+    vprint = lambda *a, **k: None #do-nothing function
+
 # Get import and export folder locations
 dom = parse(ardourFile)
 sessionName = dom.getElementsByTagName("Session")[0].getAttribute("name")
@@ -26,7 +33,7 @@ sessionName = dom.getElementsByTagName("Session")[0].getAttribute("name")
 dir = os.path.dirname(ardourFile)
 importFolder = os.path.join(dir,"interchange",sessionName,"midifiles")
 exportFolder = os.path.join(dir,"export")
-print(importFolder,exportFolder)
+vprint(importFolder,exportFolder)
 
 # Iterate through the MIDI tracks in Ardour (called "Routes" in the XML file)
 # Gets Ardour track id's and saves track names
@@ -42,7 +49,7 @@ for route in dom.getElementsByTagName("Route"):
         trackRef[route.getAttribute("id")] = i
         i = i+1
 
-print(trackRef)
+vprint(trackRef)
 
 # Iterate through the MIDI source files referenced by Ardour
 # Gets file names and file id's
@@ -51,7 +58,7 @@ for source in dom.getElementsByTagName("Source"):
     if source.getAttribute("type") == "midi":
         sourceRef[source.getAttribute("id")] = source.getAttribute("name")
 
-print(sourceRef)
+vprint(sourceRef)
 
 def convertToMeasuresAndBeats(trackTotalTime):
     #beatsPerSecond = 112
@@ -65,11 +72,11 @@ def convertToMeasuresAndBeats(trackTotalTime):
 for playlist in dom.getElementsByTagName("Playlist"):
     if playlist.getAttribute("type") == "midi":
         trackNum = trackRef[playlist.getAttribute("orig-track-id")]
-        print('\n\nTrack',trackNum)
+        vprint('\n\nTrack',trackNum)
         trackTotalTime = 0 #for tracking message times between regions
         previousRegionEndTime = 0
         for region in playlist.getElementsByTagName("Region"):
-            print(os.path.join(importFolder,sourceRef[region.getAttribute("source-0")]))
+            vprint(os.path.join(importFolder,sourceRef[region.getAttribute("source-0")]))
             sourceMidi = MidiFile(os.path.join(importFolder,sourceRef[region.getAttribute("source-0")]))
             startBeats = float(region.getAttribute("start-beats")) #beat that the MIDI source file starts on for this region
             lengthBeats = float(region.getAttribute("length-beats")) #lengh of the region in beats
@@ -80,28 +87,26 @@ for playlist in dom.getElementsByTagName("Playlist"):
             for msg in sourceMidi.tracks[0]:
                 if not msg.is_meta:
                     sourceMidiTotalTime += msg.time
-
-                    #print('end track at tick:',lengthBeats*19200+startBeats*19200)
                     firstTime = int((trackTotalTime-previousRegionEndTime) + (sourceMidiTotalTime-startBeats*19200))
-                    if sourceMidiTotalTime >= startBeats*19200 and sourceMidiTotalTime <= (lengthBeats+startBeats)*19200 and firstTime >= 0:
+                    if sourceMidiTotalTime >= startBeats*19200 and sourceMidiTotalTime <= (lengthBeats+startBeats)*19200:
+                        if firstTime < 0:
+                            print("\n\nERROR: Overlapping regions on Track",mid.tracks[trackNum].name,convertToMeasuresAndBeats(firstTime+previousRegionEndTime),'\n\n\n\n')
                         if firstPass:
-                            print('trackTotalTime',convertToMeasuresAndBeats(trackTotalTime))
-                            print('sourceMidiTotalTime',convertToMeasuresAndBeats(sourceMidiTotalTime))
-                            print('startBeats',convertToMeasuresAndBeats(startBeats))
+                            vprint('trackTotalTime',convertToMeasuresAndBeats(trackTotalTime))
+                            vprint('sourceMidiTotalTime',convertToMeasuresAndBeats(sourceMidiTotalTime))
+                            vprint('startBeats',convertToMeasuresAndBeats(startBeats))
                             firstTime = int((trackTotalTime-previousRegionEndTime) + (sourceMidiTotalTime-startBeats*19200))
-                            #print('position',position,'+ totalTime',totalTime,'- start',start,'=',firstTime,convertToMeasuresAndBeats(firstTime))
                             editedMsg = msg.copy(time=firstTime)
                             trackTotalTime = firstTime+previousRegionEndTime
                             firstPass = False
-                            #if (firstTime<0):
-                            #    print("CAN'T HAVE NEGATIVE\n\n\n")
                         else:
                             editedMsg = msg.copy(time=msg.time)
                             trackTotalTime += msg.time
-                        print(editedMsg,convertToMeasuresAndBeats(trackTotalTime))
+                        vprint(editedMsg,convertToMeasuresAndBeats(trackTotalTime))
                         mid.tracks[trackNum].append(editedMsg)
             previousRegionEndTime = trackTotalTime
 
 
-mid.save(os.path.join(exportFolder,"ArdourMidiExport.mid"))
-print("\nFile saved to your Ardour file's export folder")
+exportFile = os.path.join(exportFolder,"ArdourMidiExport.mid")
+mid.save(exportFile)
+print("\nMIDI exported\n",exportFile)