+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Change log
 
-## next
+## 0.2
 * fixed import of tracks that start later in piece rather than at very beginning
+* fixed import of tracks with multiple regions (regions need to be non-overlapping)
 * changed program number to match instrument (partial support)
 * added command line argument for file path
+* added command line argument for verbose output
 
 ## 0.1
 * initial MIDI export